subtree updates
poky: da0ce760c5..14c5392fde:
  Andrej Valek (1):
        busybox: add tmpdir option into mktemp applet

  Anuj Mittal (2):
        documentation: prepare for 3.3.2 release
        poky.conf: bump version for 3.3.2 hardknott release

  Asfak Rahman (1):
        openssh: Remove temporary keys before generating new ones

  Bruce Ashfield (6):
        linux-yocto/5.10: update to v5.10.47
        linux-yocto/5.4: update to v5.4.129
        linux-yocto/5.10: scsi-debug needs scsi-disk
        linux-yocto-dev: base AUTOREV on specified version
        kernel-devsrc: fix scripts/prepare for ARM64
        kernel-devsrc: fix scripts prepare for powerpc

  Changqing Li (2):
        libconvert-asn1-perl: fix CVE-2013-7488
        boost-build-native: workaround one rarely hang problem on fedora34

  Florian Amstutz (1):
        devtool: deploy-target: Fix preserving attributes when using --strip

  Kai Kang (1):
        rxvt-unicode: fix CVE-2021-33477

  Khairul Rohaizzat Jamaluddin (2):
        curl: Fix CVE-2021-22898
        curl: Fix CVE-2021-22897

  Marek Vasut (1):
        linux-firmware: Package RSI 911x WiFi firmware

  Mingli Yu (1):
        perl: correct libpth and glibpth

  Richard Purdie (9):
        oeqa/selftest/runcmd: Tweal test timeouts
        sstate/staging: Handle directory creation race issue
        oeqa/selftest/archiver: Allow tests to ignore empty directories
        runqemu: Remove potential lock races around tap device handling
        glibc-testsuite: Fix build failures when directly running recipe
        oeqa/selftest/multiprocesslauch: Fix test race
        dwarfsrcfiles: Avoid races over debug-link files
        bitbake: data_smart/parse: Allow ':' characters in variable/function names
        bitbake: data_smart: Allow colon in variable expansion regex

  Vinay Kumar (1):
        binutils: Fix CVE-2021-20197

  Wadim Egorov (1):
        xserver-xorg: Fix builds without glx

  wangmy (2):
        go: upgrade 1.16.3 -> 1.16.4
        go: upgrade 1.16.4 -> 1.16.5

  zhengruoqin (1):
        busybox: upgrade 1.33.0 -> 1.33.1

meta-raspberrypi: 064f5404ea..9d372828ba:
  Martin Jansa (1):
        python3-adafruit-*: fix branch parameter

meta-openembedded: c51e79dd85..5a4b2ab29d:
  Adrian Zaharia (1):
        ntp: fix ntpdate to wait for subprocesses

  Akifumi Chikazawa (1):
        openvpn: add CVE-2020-7224 and CVE-2020-27569 to allowlist

  Andreas Müller (1):
        mariadb: Fix configure

  Armin Kuster (1):
        hiawatha: fix url.

  Changqing Li (1):
        nginx: fix CVE-2021-23017

  Gianfranco (3):
        vboxguestdrivers: upgrade 6.1.18 -> 6.1.20
        vboxguestdrivers: upgrade 6.1.20 -> 6.1.22
        vboxguestdrivers: add a fix for build failure with kernel 5.13

  Joe Slater (1):
        python3-pillow: fix CVE-2021-34552

  Kai Kang (1):
        mariadb: fix failures to start install_db.service

  Khem Raj (3):
        mariadb: Upgrade to 10.5.10
        mariadb: Include missing sys/type.h for ssize_t
        mariadb: Fix build with clang/musl

  Leon Anavi (1):
        python3-urllib3: Upgrade 1.26.4 -> 1.26.5

  Li Wang (1):
        apache2: fix CVE-2020-13950 CVE-2020-35452 CVE-2021-26690 CVE-2021-26691 CVE-2021-30641

  Masaki Ambai (1):
        nss: add CVE-2006-5201 to allowlist

  Sam Van Den Berge (1):
        libiio: fix installing libiio when python3 bindings are enabled

  Tony Tascioglu (2):
        redis: fix CVE-2021-29477
        redis: fix CVE-2021-29478

  Trevor Gamblin (1):
        python3-django: upgrade 3.2.4 -> 3.2.5 (fix CVE-2021-35042)

  Zoltán Böszörményi (2):
        mariadb: Use qemu to run cross-compiled binaries
        mariadb: Upgrade to 10.5.11

  massimo toscanelli (1):
        sysbench: fix memory test

+++ poky/meta/classes/sstate.bbclass
@@ -483,7 +483,7 @@ def sstate_clean_cachefiles(d):
         ss = sstate_state_fromvars(ld, task)
         sstate_clean_cachefile(ss, ld)
 
-def sstate_clean_manifest(manifest, d, prefix=None):
+def sstate_clean_manifest(manifest, d, canrace=False, prefix=None):
     import oe.path
 
     mfile = open(manifest)
@@ -501,7 +501,9 @@ def sstate_clean_manifest(manifest, d, prefix=None):
             if entry.endswith("/"):
                 if os.path.islink(entry[:-1]):
                     os.remove(entry[:-1])
-                elif os.path.exists(entry) and len(os.listdir(entry)) == 0:
+                elif os.path.exists(entry) and len(os.listdir(entry)) == 0 and not canrace:
+                    # Removing directories whilst builds are in progress exposes a race. Only
+                    # do it in contexts where it is safe to do so.
                     os.rmdir(entry[:-1])
             else:
                 os.remove(entry)
@@ -539,7 +541,7 @@ def sstate_clean(ss, d):
         for lock in ss['lockfiles']:
             locks.append(bb.utils.lockfile(lock))
 
-        sstate_clean_manifest(manifest, d)
+        sstate_clean_manifest(manifest, d, canrace=True)
 
         for lock in locks:
             bb.utils.unlockfile(lock)
